Add if context for each function

+++ setup.sh
@@ -105,28 +105,36 @@ setup() {
     }
 
     init_node() {
-        curl -L git.io/nodebrew | perl - setup
+        if ! which nodebrew; then
+            curl -L git.io/nodebrew | perl - setup
+        fi
     }
 
     init_yarn() {
-        curl -sS https://dl.yarnpkg.com/debian/pubkey.gpg | sudo apt-key add -
-        echo "deb https://dl.yarnpkg.com/debian/ stable main" | sudo tee /etc/apt/sources.list.d/yarn.list
-        sudo apt-get update && sudo apt-get -y install yarn
+        if ! which yarn; then
+            curl -sS https://dl.yarnpkg.com/debian/pubkey.gpg | sudo apt-key add -
+            echo "deb https://dl.yarnpkg.com/debian/ stable main" | sudo tee /etc/apt/sources.list.d/yarn.list
+            sudo apt-get update && sudo apt-get -y install yarn
+        fi
     }
 
     init_docker() {
-        sudo apt-get remove docker docker-engine docker.io containerd runc
-        sudo apt-get install apt-transport-https ca-certificates curl gnupg-agent software-properties-common
-        curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
-        sudo add-apt-repository "deb [arch=amd64] https://download.docker.com/linux/ubuntu $(lsb_release -cs) stable"
-        sudo apt-get update
-        sudo apt-get install docker-ce docker-ce-cli containerd.io
-        sudo usermod -aG docker $USER
+        if ! which docker; then
+            sudo apt-get remove docker docker-engine docker.io containerd runc
+            sudo apt-get install apt-transport-https ca-certificates curl gnupg-agent software-properties-common
+            curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
+            sudo add-apt-repository "deb [arch=amd64] https://download.docker.com/linux/ubuntu $(lsb_release -cs) stable"
+            sudo apt-get update
+            sudo apt-get install docker-ce docker-ce-cli containerd.io
+            sudo usermod -aG docker $USER
+        fi
     }
 
     init_docker_compose() {
-        sudo curl -L https://github.com/docker/compose/releases/download/1.21.1/docker-compose-`uname -s`-`uname -m` -o /usr/local/bin/docker-compose
-        sudo chmod 0755 /usr/local/bin/docker-compose
+        if ! which docker-compose; then
+            sudo curl -L https://github.com/docker/compose/releases/download/1.21.1/docker-compose-`uname -s`-`uname -m` -o /usr/local/bin/docker-compose
+            sudo chmod 0755 /usr/local/bin/docker-compose
+        fi
     }
 
     init_git
